add some examples

+++ src/main/java/service/Calculator.java
@@ -4,7 +4,7 @@ public interface Calculator {
 
     int sum(int a, int b);
 
-    int div(int a, int b);
+    double div(int a, int b);
 
     double logByBase(double base, double value);
